Separate QA module for heap attack tests
