+++ x-pack/plugin/esql/qa/server/heap-attack/build.gradle
@@ -6,7 +6,7 @@ dependencies {
 
 restResources {
   restApi {
-    include '_common', 'bulk', 'indices', 'esql', 'xpack'
+    include '_common', 'bulk', 'indices', 'esql', 'xpack', 'enrich'
   }
 }
